requirements.txt
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.13.0
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==3.2.1
defusedxml==0.6.0
diff-match-patch==20200713
DingtalkChatbot==1.5.1
distlib==0.3.1
Django==2.2.16
django-bootstrap4==2.3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.4.0
django-oss-storage==1.1.1
django-pure-pagination==0.3.0
django-python3-ldap==0.11.3
django-ranged-response==0.2.0
django-registration-redux==2.8
django-rest-framework==0.1.0
django-reversion==3.0.8
django-simple-captcha==0.5.13
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.3.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
ldap3==2.8.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
oss2==2.13.1
Pillow==8.0.1
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
sentry-sdk==0.19.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
utils==1.0.1
virtualenv==20.2.0
virtualenvwrapper-win==1.2.6
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==1.2.0
xlwt==1.3.0
yunpian==0.1.1
zipp==3.4.0
